Added fast images feature
With fast images, you can cache images much faster than just letting them archive themselves with NSCoding.

+++ Pod/Classes/CKSQLiteCache.m
@@ -184,7 +184,7 @@
 		NSLog(@"Error reading file attributes: %@", error);
 	}
 	
-	return attributes.fileSize;
+	return (NSUInteger)attributes.fileSize;
 }
 
 @end
